refactor: move form logic to Form component
