+++ src/components/Form/Form.js
@@ -1,14 +1,56 @@
-import { useEffect } from 'react';
+import { useEffect, useCallback, useState } from 'react';
 import './Form.css';
 
 import Button from '../Button/Button';
 import TextInput from '../TextInput/TextInput';
 
-function Form({ modalType, onClose, isOpen, isValid, handleChange, errors, values, onSubmit, showErrors }) {
+import { errorMessages } from '../../config/form';
+
+function Form({ modalType, isOpen, setIsOpen }) {
+  const [values, setValues] = useState({});
+  const [errors, setErrors] = useState(errorMessages);
+  const [showErrors, setShowErrors] = useState(false);
+  const [isValid, setIsValid] = useState(false);
+
+  const resetForm = useCallback(
+    (newValues = { email: '', password: '', message: '' }, newErrors = {}, newIsValid = false) => {
+      setValues(newValues);
+      setErrors(newErrors);
+      setIsValid(newIsValid);
+      setShowErrors(false);
+    },
+    [setValues, setErrors, setIsValid]
+  );
+
+  const handleChange = (event) => {
+    const { target } = event;
+    const { name, value } = target;
+    setValues({ ...values, [name]: value });
+    setErrors({ ...errors, [name]: target.validationMessage || '' });
+    setIsValid(target.closest('form').checkValidity());
+  };
+
+  const closeModal = (evt) => {
+    evt.stopPropagation();
+    if (evt.type === 'click' || evt.key === 'Escape') {
+      setIsOpen(false);
+      resetForm();
+    }
+  };
+
+  const handleSubmit = (event, canSubmit) => {
+    event.stopPropagation();
+    if (canSubmit) {
+      closeModal(event);
+    } else {
+      setShowErrors(true);
+    }
+  };
+
   useEffect(() => {
-    document.addEventListener('keydown', onClose);
+    document.addEventListener('keydown', closeModal);
     return () => {
-      document.removeEventListener('keydown', onClose);
+      document.removeEventListener('keydown', closeModal);
     };
   });
   return (
@@ -21,7 +63,7 @@ function Form({ modalType, onClose, isOpen, isValid, handleChange, errors, value
         noValidate
       >
         <h2 className="form__title">Write us</h2>
-        <button className="form__close-button clickable" type="button" aria-label="close-modal" onClick={onClose} />
+        <button className="form__close-button clickable" type="button" aria-label="close-modal" onClick={closeModal} />
         <TextInput
           values={values}
           errors={errors}
@@ -49,15 +91,15 @@ function Form({ modalType, onClose, isOpen, isValid, handleChange, errors, value
           required
           onChange={handleChange}
         />
-        <Button onClick={(event) => onSubmit(event, isValid)} label="Submit" location="form" isValid={isValid} />
+        <Button onClick={(event) => handleSubmit(event, isValid)} label="Submit" location="form" isValid={isValid} />
       </form>
       <div
         role="button"
         aria-label="close-modal"
         tabIndex={0}
         className={`overlay ${isOpen ? 'overlay_visible' : ''}`}
-        onClick={onClose}
-        onKeyDown={onClose}
+        onClick={closeModal}
+        onKeyDown={closeModal}
       />
     </>
   );
